MRD-2698 Move Prison API classes
