+++ src/test/kotlin/uk/gov/justice/digital/hmpps/makerecalldecisionapi/client/prisonapi/PrisonApiClientTest.kt
@@ -18,7 +18,6 @@ import org.springframework.http.HttpStatus.NOT_FOUND
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.EndpointMocker.Companion.mockGetEndpointWithFailure
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.EndpointMocker.Companion.mockGetEndpointWithSuccess
-import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.PrisonApiClient
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.prisonapi.domain.prisonApiOffenderMovement
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.prisonapi.domain.toJsonString
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.exception.ClientTimeoutRuntimeException
