+++ src/test/kotlin/uk/gov/justice/digital/hmpps/makerecalldecisionapi/integration/client/PrisonApiClientTest.kt
@@ -9,7 +9,7 @@ import org.mockserver.model.HttpResponse
 import org.mockserver.model.MediaType
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.ActiveProfiles
-import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.PrisonApiClient
+import uk.gov.justice.digital.hmpps.makerecalldecisionapi.client.prisonapi.PrisonApiClient
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.exception.NotFoundException
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.integration.IntegrationTestBase
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.integration.responses.prison.agencyResponse
